+++ Changelog.md
@@ -2,12 +2,12 @@
 
 ## Common
 
-- Add `markdown_summary` to `simulation_manager`
+- Add `markdown_summary`
 
 ## CLI
 
 - Call `markdown_summary` at the end and print to stdout
-- Add `--markdown` argument to print to file
+- Change `--summary` argument to print to file
 
 # 2.2.3